+++ OUDS/Core/ComponentsUIKit/Sources/Controls/Radio/_OUDSRadio.swift
@@ -17,7 +17,7 @@ import OUDSFoundations // NOTE: "unused" false positive by periphery (https://gi
 import SwiftUI
 import UIKit
 
-// MARK: - OUDSThemesContract Radio View Controller
+// MARK: - OUDS Radio View Controller
 
 /// UIKit `UIViewController`  hosting view controllers so as to expose the SwiftUI `OUDSRadio`
 /// Helps to manage the states and values of the `OUDSRadio` and tries to expose a UIKit-like API for actions.
@@ -136,7 +136,7 @@ public final class OUDSRadioViewController: UIViewController {
     }
 }
 
-// MARK: - OUDSThemesContract Radio Wrapper
+// MARK: - OUDS Radio Wrapper
 
 /// A SwiftUI `View` which embeds the SwiftUI `OUDSRadio` and exposes bindings
 /// and view model.
@@ -162,7 +162,7 @@ struct OUDSRadioWrapper: View {
     }
 }
 
-// MARK: - OUDSThemesContract Radio View Model
+// MARK: - OUDS Radio View Model
 
 /// The `SwiftUI` view model used inside the `OUDSRadioWrapper` to manage the state of the embeded `OUDSRadio`
 @MainActor final class OUDSRadioViewModel: ObservableObject {
@@ -188,7 +188,7 @@ struct OUDSRadioWrapper: View {
     deinit {}
 }
 
-// MARK: - OUDSThemesContract SwiftUI Bridge
+// MARK: - OUDS SwiftUI Bridge
 
 extension OUDSUIKitBrige {