+++ OUDS/Core/ThemesContract/Sources/_ThemesContract.docc/Tutorial/swift-samples/tutorial-4-4.swift
@@ -31,7 +31,7 @@ struct ContentView: View {
         VStack {
             HStack {
                 OUDSTag(label: "Welcome")
-                OUDSTag(label: "OUDSThemesContract")
+                OUDSTag(label: "OUDS")
                 OUDSTag(label: "Tutorial")
             }